Move CameraVisPatch-related functions into CameraVisPatch
