+++ content_scripts/CameraVis.js
@@ -10,19 +10,6 @@ import {
 } from './Shaders.js';
 import {VisualDiff} from './VisualDiff.js';
 
-/**
- * All data serialized to store a CameraVis patch (3d picture)
- * Notably `key` corresponds to frame key
- * @typedef {{
- *   key: string,
- *   container: Array<number>,
- *   phone: Array<number>,
- *   texture: string,
- *   textureDepth: string,
- *   creationTime: number,
- * }} PatchSerialization
- */
-
 const debug = false;
 
 function setMatrixFromArray(matrix, array) {
@@ -163,7 +150,7 @@ export class CameraVis {
 
         return {
             key: frameKey,
-            patch: CameraVis.createPatch(
+            patch: CameraVisPatch.createPatch(
                 this.container.matrix,
                 this.phone.matrix,
                 this.texture.image,
@@ -173,108 +160,7 @@ export class CameraVis {
         };
     }
 
-    /**
-     * @param {PatchSerialization} serialization
-     * @return {string} frame key
-     */
-    createToolForPatchSerialization(serialization) {
-        let toolMatrix = new THREE.Matrix4().fromArray(serialization.phone);
-        let containerMatrix = new THREE.Matrix4().fromArray(serialization.container);
-        containerMatrix.elements[13] = 0;
-        toolMatrix.premultiply(containerMatrix);
-        toolMatrix.multiply(new THREE.Matrix4().makeRotationFromEuler(new THREE.Euler(0, 0, Math.PI / 2)));
-        // let toolRotation = new THREE.Matrix4().makeRotationFromEuler(new THREE.Euler(Math.PI / 2, Math.PI / 2, 0));
-        // toolMatrix.premultiply(toolRotation);
 
-        let addedTool = realityEditor.gui.pocket.createFrame('spatialPatch', {
-            noUserInteraction: true,
-            initialMatrix: toolMatrix.elements,
-            onUploadComplete: () => {
-                realityEditor.network.postVehiclePosition(addedTool);
-                write();
-            },
-        });
-
-        const frameKey = addedTool.uuid;
-        serialization.key = frameKey;
-        const write = () => {
-            realityEditor.network.realtime.writePublicData(
-                addedTool.objectId, frameKey, frameKey + 'storage',
-                'serialization', serialization
-            );
-        };
-        setTimeout(write, 500);
-        setTimeout(write, 3000);
-
-        return addedTool.uuid;
-    }
-
-    /**
-     * @param {Array<number>} containerMatrix - array representing 4x4 matrix from threejs
-     * @param {Array<number>} phoneMatrix - array representing 4x4 matrix from threejs
-     * @param {string} textureImage - base64 data url for texture
-     * @param {string} textureDepthImage - base64 data url for depth texture
-     * @param {number} creationTime - Time when patch created. Usually from Date.now()
-     * @return {CameraVisPatch}
-     */
-    static createPatch(containerMatrix, phoneMatrix, textureImage, textureDepthImage, creationTime) {
-        let patch = new THREE.Group();
-        patch.matrix.copy(containerMatrix);
-        patch.matrixAutoUpdate = false;
-        patch.matrixWorldNeedsUpdate = true;
-
-        let phone = new THREE.Group();
-        phone.matrix.copy(phoneMatrix);
-        phone.matrixAutoUpdate = false;
-        phone.matrixWorldNeedsUpdate = true;
-        phone.frustumCulled = false;
-
-        let texture = new THREE.Texture();
-        // texture.minFilter = THREE.NearestFilter;
-        // texture.magFilter = THREE.NearestFilter;
-        // texture.minFilter = THREE.LinearFilter;
-        // texture.magFilter = THREE.LinearFilter;
-        // texture.generateMipmaps = false;
-
-        let textureDepth = new THREE.Texture();
-        // textureDepth.minFilter = THREE.NearestFilter;
-        // textureDepth.magFilter = THREE.NearestFilter;
-        // textureDepth.minFilter = THREE.LinearFilter;
-        // textureDepth.magFilter = THREE.LinearFilter;
-        // textureDepth.generateMipmaps = false;
-
-        texture.image = textureImage;
-        textureDepth.image = textureDepthImage;
-
-        texture.needsUpdate = true;
-        textureDepth.needsUpdate = true;
-
-        let mesh = createPointCloud(texture, textureDepth, ShaderMode.SOLID);
-        mesh.material.uniforms.patchLoading.value = 0;
-
-        let lastTime = -1;
-        function patchLoading(time) {
-            if (lastTime < 0) {
-                lastTime = time;
-            }
-            // limit to 30fps
-            let dt = Math.min(time - lastTime, 67);
-            lastTime = time;
-            mesh.material.uniforms.patchLoading.value += 8 * dt / 1000;
-            if (mesh.material.uniforms.patchLoading.value < 1) {
-                window.requestAnimationFrame(patchLoading);
-            } else {
-                mesh.material.uniforms.patchLoading.value = 1;
-            }
-        }
-        window.requestAnimationFrame(patchLoading);
-
-        phone.add(mesh);
-        patch.add(phone);
-
-        patch.__creationTime = creationTime;
-        return new CameraVisPatch(patch, mesh, phone);
-    }
 
     setupDebugCubes() {
         let debugDepth = new THREE.MeshBasicMaterial({
